requirements.txt
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.5.0
numpy>=1.19.0
pandas>=1.1.0
matplotlib>=3.3.0
seaborn>=0.11.0
pillow>=8.0.0
albumentations>=1.0.0
scikit-image>=0.19.0
tqdm>=4.60.0
PyYAML>=5.4.0
nvidia-ml-py>=12.0.0 